# api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: CasaOS App Management API
  version: v2
  description: |
    CasaOS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.

    For issues, please visit the [GitHub repository](https://github.com/ReCasaOS/CasaOS-AppManagement).
servers:
  - url: /v2/app_management
...
security:
  - access_token: []
paths:
...
  /compose/updates:
    get:
      summary: What updating every app at once would do
      description: |
        The apps with an update on offer, and for each the images an update would
        write: the catalogue's image where the catalogue moved one (`catalogue`),
        or every image as it is where the last image check found a newer build
        under a tag the app already names (`image`). Nothing is changed by asking.

        With `check=true` the same sweep as `POST /image-updates` runs first, one
        round trip per distinct image, so the plan is as fresh as the registries;
        a registry that does not answer is not fatal.
      operationId: composeAppUpdatePlan
      tags:
        - Compose methods
      parameters:
        - name: check
          in: query
          description: Ask every image's registry first.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppUpdatePlanOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
    post:
      summary: Update the apps listed, one after another
      description: |
        The same update the button on each app runs, for every app listed, in the
        order given, one at a time, on this box rather than from the browser: a tab
        closed halfway does not stop it. Each app publishes the events an update
        publishes. The run is answered as it starts; `GET /compose/updates/run`
        says what became of each app.

        One run at a time: a second list while one is going is refused with 409.
      operationId: updateComposeApps
      tags:
        - Compose methods
      requestBody:
        $ref: "#/components/requestBodies/RequestComposeAppUpdates"
      responses:
        "202":
          $ref: "#/components/responses/ComposeAppUpdateRunOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "409":
          $ref: "#/components/responses/ResponseConflict"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /compose/updates/run:
    get:
      summary: The update of every app in progress, or the last one
      description: |
        Each app of the run and what became of it: queued, updating, updated,
        current (nothing to do, with the reason when there is one), or failed with
        the error. `finished_at` is set once the last app is done. Kept in memory:
        a restart of this service forgets the last run.
      operationId: composeAppUpdateRun
      tags:
        - Compose methods
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppUpdateRunOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
  /image-updates:
    post:
      summary: Check every installed app for a newer image
      description: |
        Asks each image's registry for the digest its tag points at now and compares it
        with the digest of the copy on disk. This sees apps that came from no app store,
        which the store's own upgradable list cannot.

        One network round trip per distinct image, so this is a deliberate action rather
        than something the app grid does on every load. The result is cached and served
        as `update_available` on each grid item until the next call.

        An image no answer can be obtained for -- an unreachable registry, credentials
        this host does not have, an image never pulled -- is listed in `unchecked`
        instead of being guessed at, and its app keeps the answer it had.
      operationId: checkImageUpdates
      tags:
        - App methods
      responses:
        "200":
          $ref: "#/components/responses/ImageUpdateCheckOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
  /apps/{id}/compose:
    get:
      summary: Get compose details of an app from registered app stores
      description: |
        Accept both `application/json` and `application/yaml` content type for the `Accept` header in the HTTP request.

        If the `Accept` header is set to `application/yaml`, the response will be the original `docker-compose.yml` content that follows [Compose Specification](https://compose-spec.io).
      operationId: composeApp
      tags:
        - AppStore methods
      parameters:
        - $ref: "#/components/parameters/StoreAppIDString"
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
...
    StoreAppIDString:
      name: id
      description: Store app ID of a compose app (unique across all app stores)
      in: path
      required: true
      schema:
        type: string
        example: "syncthing"
...
  requestBodies:
    RequestComposeAppUpdates:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ComposeAppUpdateRequest"
...
  responses:
...
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
    ResponseBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Bad Request"
    ResponseConflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Conflict"
...
    ComposeAppUpdatePlanOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppUpdatePlan"
    ComposeAppUpdateRunOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppUpdateRun"
    ImageUpdateCheckOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ImageUpdateCheckResult"
...
    ComposeAppOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppWithStoreInfo"
        application/yaml:
          schema:
            type: string
...
  schemas:
    BaseResponse:
      properties:
        message:
          description: message returned by server side if there is any
          type: string
          example: ""
...
    ComposeAppStoreInfo:
      required:
        - author
        - category
        - description
        - developer
        - icon
        - screenshot_link
        - tagline
        - thumbnail
        - title
        - tips
        - index
        - port_map
      properties:
        store_app_id:
          $ref: "#/components/schemas/StoreAppID"
        title:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Syncthing
        image:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: syncthing:latest
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Syncthing is a continuous file synchronization program. It synchronizes files between two or more computers in real time, safely protected from prying eyes. Your data is your data alone and you deserve to choose where it is stored, whether it is shared with some third party, and how it's transmitted over the internet.
        tagline:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Free, secure, and distributed file synchronisation tool.
        icon:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Syncthing/icon.png
        screenshot_link:
          type: array
          items:
            type: string
          example:
            - https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/OpenSpeedTest/screenshot-1.png
            - https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/OpenSpeedTest/screenshot-2.png
          x-oapi-codegen-extra-tags:
            mapstructure: "screenshot_link"
            yaml: "screenshot_link,omitempty"
        thumbnail:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Jellyfin/thumbnail.jpg
        author:
          type: string
          example: CasaOS Team
        developer:
          type: string
          example: Syncthing
        category:
          type: string
          example: Backup
        tips:
          $ref: "#/components/schemas/TipsStoreInfo"
        scheme:
          $ref: "#/components/schemas/Scheme"
        hostname:
          type: string
          example: ""
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        port_map:
          type: string
          example: "8384"
          x-oapi-codegen-extra-tags:
            mapstructure: "port_map"
            yaml: "port_map,omitempty"
        index:
          type: string
          example: "/"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        apps:
          title: store info of container apps included in the compose app
          description: |
            The key is the name of a container app, and the value is the StoreInfo of this container app.
          readOnly: true
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AppStoreInfo"
          example:
            syncthing: {}
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        main:
          title: name of the main service in `services`
          description: |
            The main service is one of the services that the user will interact with.
          type: string
          example: "syncthing"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        architectures:
          title: architectures supported by the compose app
          description: |
            The architectures supported by the compose app. If the architecture of the host is not in this list, the compose app will not be installed.
          type: array
          items:
            type: string
          example:
            - "amd64"
            - "arm64"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        is_uncontrolled: # uncontrolled == not stable and latest
          title: is the app uncontrolled. 
          description: |
            the mean the image tag is custom. not stable or latest
          type: boolean
          example: false
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
    ComposeApp:
      type: object
      description: |-
        See [Compose Specification](https://compose-spec.io) for the schema structure of `ComposeApp`.

      x-go-type: types.Project
      x-go-type-import:
        name: types
        path: github.com/compose-spec/compose-go/v2/types
    ComposeAppWithStoreInfo:
      description: |-
        Schema for `ComposeAppOK` response.

        > Do not mistake it as `ComposeAppStoreInfo` which is just store information of a compose app.
      properties:
        store_info:
          $ref: "#/components/schemas/ComposeAppStoreInfo"
        compose:
          $ref: "#/components/schemas/ComposeApp"
        status:
          type: string
          readOnly: true
          example: running
        update_available:
          readOnly: true
          type: boolean
          example: true
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        is_uncontrolled:
          readOnly: true
          title: is the app uncontrolled. 
          description: |
            the mean the image tag is custom. not stable or latest
          type: boolean
          example: false
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
...
    AppStoreInfo:
      title: StoreInfo of an app
      description: |-
        > Do not mistake it as *information of AppStore* which does not exist.
      required:
        - image
        - envs
        - ports
        - volumes
        - devices
      properties:
        image:
          type: string
          example: "syncthing:latest"
        envs:
          type: array
          items:
            $ref: "#/components/schemas/EnvStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        ports:
          type: array
          items:
            $ref: "#/components/schemas/PortStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        volumes:
          type: array
          items:
            $ref: "#/components/schemas/VolumeStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
...
    EnvStoreInfo:
      required:
        - container
        - description
      properties:
        container:
          type: string
          example: TZ
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Time Zone
    PortStoreInfo:
      required:
        - container
        - description
      properties:
        container:
          type: string
          example: 8384
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Web UI
...
    VolumeStoreInfo:
      required:
        - container
        - description
      properties:
        container:
          type: string
          example: /config
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Syncthing config directory
    DeviceStoreInfo:
      properties:
        container:
          type: string
          example: /dev/dri
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: GPU device
    TipsStoreInfo:
      required:
        - before_install
      properties:
        before_install:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: |
              Tips for users before install.
          x-oapi-codegen-extra-tags:
            mapstructure: "before_install"
            yaml: "before_install,omitempty"
        custom:
          type: string
          example: "custom tips by user"
    Scheme:
      type: string
      default: http
      enum:
        - http
        - https
      x-oapi-codegen-extra-tags:
        yaml: ",omitempty"
    StoreAppID:
      title: Store app ID
      description: |
        Store app ID of the compose app (unique across all app stores)

        > The `store_app_id` might be same as app name most of the time. However app name might change, due to cloning for example.
        > If app name has changed, there is no way to associate it with the original app in the app store, unless we use the `store_app_id`.
      readOnly: true
      type: string
      example: "syncthing"
      x-go-name: StoreAppID
      x-oapi-codegen-extra-tags:
        mapstructure: "store_app_id"
        yaml: "store_app_id,omitempty"
...
    ComposeAppUpdatePlan:
      description: What updating every app at once would do.
      required:
        - apps
      properties:
        apps:
          type: array
          items:
            $ref: "#/components/schemas/ComposeAppUpdatePlanApp"
    ComposeAppUpdatePlanApp:
      required:
        - id
        - title
        - icon
        - kind
        - services
      properties:
        id:
          type: string
          example: jellyfin
        title:
          type: string
          example: Jellyfin
        icon:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Jellyfin/icon.png
        kind:
          description: |
            `catalogue` when the catalogue moves at least one image, which is then
            listed with the image it moves to; `image` when the images stay named
            as they are and the last image check found a newer build under one of
            them, in which case every service is listed as it is.
          type: string
          enum:
            - catalogue
            - image
        services:
          type: array
          items:
            $ref: "#/components/schemas/ComposeAppUpdatePlanService"
    ComposeAppUpdatePlanService:
      required:
        - name
        - from
        - to
      properties:
        name:
          type: string
          example: jellyfin
        from:
          type: string
          example: jellyfin/jellyfin:10.9.11
        to:
          type: string
          example: jellyfin/jellyfin:10.10.3
    ComposeAppUpdateRequest:
      required:
        - ids
      properties:
        ids:
          description: The apps to update, in this order.
          type: array
          items:
            type: string
          example:
            - jellyfin
            - syncthing
    ComposeAppUpdateRun:
      description: An update of every app, in progress or done.
      required:
        - started_at
        - apps
      properties:
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        apps:
          type: array
          items:
            $ref: "#/components/schemas/ComposeAppUpdateRunApp"
    ComposeAppUpdateRunApp:
      required:
        - id
        - state
      properties:
        id:
          type: string
          example: jellyfin
        state:
          type: string
          enum:
            - queued
            - updating
            - updated
            - current
            - failed
        message:
          description: The reason, when the app stayed current for one, or the error.
          type: string
    ImageUpdateCheckResult:
      description: What one check pass found.
      required:
        - updatable
        - unchecked
      properties:
        updatable:
          description: Apps whose registry holds a different image than the copy on disk.
          type: array
          items:
            type: string
          example:
            - syncthing
        unchecked:
          description: |
            Apps no answer could be obtained for, and why. A registry that cannot be
            reached is not evidence that nothing changed, so these keep the answer they
            had rather than being reported as up to date.
          type: object
          additionalProperties:
            type: string
          example:
            plex: image never pulled on this host
...